SO MUCH PROGRESS. The Grid holds roads now.

+++ cartographer/Coordinate.java
@@ -1,19 +1,19 @@
 public class Coordinate {
-    public double x, y;
-    public Coordinate(double x, double y) {
+    public int x, y;
+    public Coordinate(int x, int y) {
         this.x = x;
         this.y = y;
     }
 
-    public double getX() {
+    public int getX() {
         return this.x;
     }
 
-    public double getY() {
+    public int getY() {
         return this.y;
     }
 
     public double distanceToCoordinate(Coordinate otherCoordinate) {
-        return Math.sqrt(Math.pow(x - otherCoordinate.getX(), 2) + Math.pow(y - otherCoordinate.getY(), 2));
+        return Math.sqrt((double)Math.pow(x - otherCoordinate.getX(), 2) + (double)Math.pow(y - otherCoordinate.getY(), 2));
     }
 }
